+++ src/js/bootstrapValidator.js
@@ -488,17 +488,7 @@
             }
 
             // Focus to the first invalid field
-            var $firstInvalidField = this.$invalidFields.eq(0);
-            if ($firstInvalidField) {
-                // Activate the tab containing the invalid field if exists
-                var $tab = $firstInvalidField.parents('.tab-pane'),
-                    tabId;
-                if ($tab && (tabId = $tab.attr('id'))) {
-                    $('a[href="#' + tabId + '"][data-toggle="tab"]').trigger('click.bs.tab.data-api');
-                }
-
-                $firstInvalidField.focus();
-            }
+            this.$invalidFields.eq(0).focus();
         },
 
         /**
@@ -752,14 +742,6 @@
                 }
             }
 
-            // Determine the tab containing the element
-            var $tabPane = $field.parents('.tab-pane'),
-                tabId,
-                $tab;
-            if ($tabPane && (tabId = $tabPane.attr('id'))) {
-                $tab = $('a[href="#' + tabId + '"][data-toggle="tab"]').parent();
-            }
-
             // Show/hide error elements and feedback icons
             $errors.attr('data-bv-result', status);
             switch (status) {
@@ -770,9 +752,6 @@
                     if ($icon) {
                         $icon.removeClass(this.options.feedbackIcons.valid).removeClass(this.options.feedbackIcons.invalid).addClass(this.options.feedbackIcons.validating).show();
                     }
-                    if ($tab) {
-                        $tab.removeClass('bv-tab-success').removeClass('bv-tab-error');
-                    }
                     break;
 
                 case this.STATUS_INVALID:
@@ -781,9 +760,6 @@
                     $parent.removeClass('has-success').addClass('has-error');
                     if ($icon) {
                         $icon.removeClass(this.options.feedbackIcons.valid).removeClass(this.options.feedbackIcons.validating).addClass(this.options.feedbackIcons.invalid).show();
-                    }
-                    if ($tab) {
-                        $tab.removeClass('bv-tab-success').addClass('bv-tab-error');
                     }
                     break;
 
@@ -802,9 +778,6 @@
                     }
 
                     $parent.removeClass('has-error has-success').addClass(this.isValidContainer($parent) ? 'has-success' : 'has-error');
-                    if ($tab) {
-                        $tab.removeClass('bv-tab-success').removeClass('bv-tab-error').addClass(this.isValidContainer($tabPane) ? 'bv-tab-success' : 'bv-tab-error');
-                    }
                     break;
 
                 case this.STATUS_NOT_VALIDATED:
@@ -814,9 +787,6 @@
                     $parent.removeClass('has-success').removeClass('has-error');
                     if ($icon) {
                         $icon.removeClass(this.options.feedbackIcons.valid).removeClass(this.options.feedbackIcons.invalid).removeClass(this.options.feedbackIcons.validating).hide();
-                    }
-                    if ($tab) {
-                        $tab.removeClass('bv-tab-success').removeClass('bv-tab-error');
                     }
                     break;
             }